Changes to base and output scan speed vars
Changed the detection scan speed supplied to be named `base_scan_speed`
and altered the logic so that `output_scan_speed` would be required for
all the layer groups but the first if layer groups are present.

Fixed bugs regarding process parameter output for both the single group
and multiple group cases.

Edited README in accordance to changes made in the f-speed detection